refactor: Split session history and channel adapters into packages
