+++ cmd/app/main.go
@@ -18,7 +18,7 @@ import (
 	"github.com/pardnchiu/agenvoy/internal/runtime"
 	"github.com/pardnchiu/agenvoy/internal/runtime/discord"
 	"github.com/pardnchiu/agenvoy/internal/runtime/telegram"
-	"github.com/pardnchiu/agenvoy/internal/session"
+	sessionHistory "github.com/pardnchiu/agenvoy/internal/session/history"
 	sessionSummary "github.com/pardnchiu/agenvoy/internal/session/summary"
 	"github.com/pardnchiu/agenvoy/internal/toolAdapter/mcp"
 )
@@ -102,7 +102,7 @@ func setSummaryCron() {
 		}
 
 		for _, sid := range sessions {
-			histories, _ := session.GetHistory(sid)
+			histories, _ := sessionHistory.Get(sid)
 			summaryHistories := summary.Get(histories)
 			if len(summaryHistories) == 0 {
 				continue
