+++ internal/runtime/discord/pending.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/pardnchiu/agenvoy/internal/filesystem"
 	"github.com/pardnchiu/agenvoy/internal/runtime"
-	sessionManager "github.com/pardnchiu/agenvoy/internal/session"
+	sessionDiscord "github.com/pardnchiu/agenvoy/internal/session/discord"
 	"github.com/pardnchiu/agenvoy/internal/utils"
 )
 
@@ -27,7 +27,7 @@ func newPendingListener(bot *Bot) *runtime.Listener[string, string] {
 }
 
 func (t *discordTransport) LookupChatID(sessionID string) (string, error) {
-	channelID, err := sessionManager.GetChannelID(sessionID)
+	channelID, err := sessionDiscord.GetChannel(sessionID)
 	if err != nil {
 		return "", fmt.Errorf("GetChannelID: %w", err)
 	}
